moved isLogingOut state to uiReducer
